pathAI/requirements.txt
gradio==4.44.1
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
plotly==5.18.0